+++ internal/config/open_unix.go
@@ -1,0 +1,9 @@
+//go:build !windows
+
+package config
+
+import "os"
+
+func openConfig(path string) (*os.File, error) {
+	return os.Open(path)
+}