+++ internal/home/config_mode_unix.go
@@ -1,0 +1,7 @@
+//go:build !windows
+
+package home
+
+func secureOptionalConfig(string) error {
+	return nil
+}